PR테스트 : 로그인페이지 잘 뜨는지 확인

See merge request s12-webmobile2-sub1/S12P11A707!13

+++ backend/ttabong/src/main/java/com/ttabong/repository/user/VolunteerRepository.java
@@ -3,6 +3,6 @@ package com.ttabong.repository.user;
 import com.ttabong.entity.user.Volunteer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VolunteerRepository extends JpaRepository<Volunteer, Integer> {
+public interface VolunteerRepository extends JpaRepository<Volunteer, Long> {
     // 봉사자 쿼리메소드 적기
 }
